Add support for Vagrant development environment
This commit adds Vagrantfile and Cheffile, which are used by Vagrant to
automatically set up a virtual development environment designed to be
used for local Bike Index development. Chef is used to automatically
provision the environment with all of the required dependencies and
configures PostgreSQL. This makes the process of setting up a local
development environment much quicker and straightforward. As long as
Vagrant/VirtualBox is installed with the required Vagrant plugins, the
environment can be brought up just by running `vagrant up`. After SSHing
in, the "Running Bike Index locally" intructions in the README can be
followed verbatim without any additional configuration.

This commit also updates README.markdown to acknowledge this feature.